finsihed subject api working on get all paged

+++ SubjectService/Database/Entities/Subject.cs
@@ -14,7 +14,7 @@ namespace SubjectService.Database.Entities
         public string Subtitle { get; set; }
         public string Content { get; set; }
         public bool IsVisible { get; set; }
-        public short Type { get; set; }
+        public int Type { get; set; }
 
     }
 }
